requirements.txt
discord.py
asyncio
youtube-dl
PyNaCl
ffmpeg
requests
BeautifulSoup